ItemMedKit.HpResource is a float

+++ Fuyu.Backend.BSG/DTO/Items/ItemMedKit.cs
@@ -6,6 +6,6 @@ namespace Fuyu.Backend.EFT.DTO.Items
     public class ItemMedKit
     {
         [DataMember]
-        public int HpResource;
+        public float HpResource;
     }
 }
